+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [13.1.3] - 2018-10-17
 ### Modified
 - Audit.WebApi and Audit.Mvc: Upgrading AspNetCore libraries from GitHub suggestion because of security vulnerability.
+- Audit.AzureDocumentDB: Adding netstandard2.0 as target
 
 ## [13.1.2] - 2018-09-12
 ### Modified